Rewrite it, use Numba, combined kernels, see notes
Rewrite it to make it more consistent with the implementations from
tradeview and others. These focus only on the main signal and smooth
signal. The histogram is trivially computed outside, but one
implementation allows the user to compute the momentum of the main and
smooth TMO signals as well.
Take advantage of convolution properties to combine the summation, a
convolution with a uniform kernel of 1s, into a convolution, but
combine it with the main EMA kernel.
If the MA mode is not EMA, we need to just use the main uniform
convolution and use the Pandas-TA moving averages.
Simplify docs